Organize and separate the code, add comments, fix some widgets, and use less code.
