+++ lib/horizontal_page.dart
@@ -3,6 +3,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'data.dart';
+import 'providers/meta_data_coran_pages_provider.dart';
+import 'providers/user_pref_provider.dart';
 
 class HorizontalPage extends ConsumerWidget {
   const HorizontalPage({super.key});
@@ -13,6 +15,7 @@ class HorizontalPage extends ConsumerWidget {
     final PageController pageController =
         PageController(initialPage: pageIndex);
 
+    // go to the user last visited page
     ref.listen<int>(
       pageIndexProvider,
       (int? previousCount, int newCount) {
@@ -23,6 +26,7 @@ class HorizontalPage extends ConsumerWidget {
       },
     );
 
+    // save the current page that the user is visiting
     Future<void> saveCurrentPage() async {
       final prefs = await SharedPreferences.getInstance();
       prefs.setInt('mushaf01_page', ref.read(pageIndexProvider));
@@ -35,13 +39,14 @@ class HorizontalPage extends ConsumerWidget {
         scrollDirection: Axis.horizontal,
         controller: pageController,
         onPageChanged: (int page) => {
+          // save the new visited page
           ref.read(pageIndexProvider.notifier).state = page,
           saveCurrentPage()
         },
-        itemBuilder: (context, index) {
-          return ref.read(pageWidgetProvider)[index];
-        },
         itemCount: 604,
+        itemBuilder: (context, index) {
+          return ref.read(coranPagesProvider)[index];
+        },
       ),
     );
   }
