+++ lib/models/page_rub_hizb_juz.dart
@@ -1,0 +1,12 @@
+class PageRubHizbJuz {
+  final int rub;
+  final int hizb;
+  final int juz;
+  final String pageInfo;
+  PageRubHizbJuz({
+    required this.rub,
+    required this.hizb,
+    required this.juz,
+    required this.pageInfo,
+  });
+}
